+++ streaming/src/main/java/com/crypto/streaming/abstracts/Exchange.java
@@ -2,4 +2,6 @@ package com.crypto.streaming.abstracts;
 
 public interface Exchange {
     public abstract void start() throws InterruptedException;
+
+    void onMessage();
 }
