+++ streaming/src/main/java/com/crypto/streaming/config/ExchangeConfig.java
@@ -1,6 +1,6 @@
 package com.crypto.streaming.config;
 
-import org.java_websocket.client.WebSocketClient;
+import com.crypto.streaming.exchange.ByBit;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -11,26 +11,11 @@ import java.util.Map;
 @Configuration
 @PropertySource("classpath:api.properties")
 public class ExchangeConfig {
-    @Value("${upbit.websocket.uri}")
-    private String upBitApiUri;
-
-    @Value("${upbit.api_key}")
-    private String upBitApiKey;
-
-    @Value("${bybit.websocket.testuri}")
-    private String byBitApiTestUri;
 
     @Bean
-    public WebSocketClient upBitWebSocketClient() {
+    public ByBit byBit(@Value("${bybit.websocket.testuri}") String uri) {
         Map<String, String> requestHeaders = new HashMap<>();
-        requestHeaders.put("Authorization", "Baerer " + upBitApiKey);
 
-        return new CryptoWebSocketClient(upBitApiUri, requestHeaders);
-    }
-
-    @Bean
-    public WebSocketClient byBitWebSocketClient() {
-        Map<String, String> requestHeaders = new HashMap<>();
-        return new CryptoWebSocketClient(byBitApiTestUri, requestHeaders);
+        return new ByBit(uri, requestHeaders);
     }
 }
