+++ streaming/src/main/resources/api.properties
@@ -1,4 +1,4 @@
 upbit.websocket.uri=wss://api.upbit.com/websocket/v1
 upbit.api_key=AKTASdOCaUPbY4VNbGyxVw1Ag1zKXRLrX7yu9DaN
 
-bybit.websocket.testuri=wss://stream-testnet.bybit.com/realtime_private
+bybit.websocket.testuri=wss://stream-testnet.bybit.com/v5/public/spot
